[LPT] AvgPool child recursive extend
