+++ inference-engine/src/transformations/src/transformations/low_precision/avg_pool.cpp
@@ -31,12 +31,12 @@ bool AvgPoolTransformation::transform(TransformationContext& context, ngraph::pa
 
     const std::shared_ptr<Node> pooling = separateInStandaloneBranch(m.get_match_root());
 
-    const auto children = pooling->get_output_target_inputs(0);
+    const std::vector<std::shared_ptr<ngraph::Node>> children = getChildrenRecursivelyExceptPrecisionPreserved(pooling);
     bool updatePrecision = false;
     // NOTE: This check was added for models that don't have FQ after AvgPool
     //       They will have transparent precision as it was in old LPT.
     for (const auto& child : children) {
-        if (!is_type<opset1::FakeQuantize>(child.get_node()->shared_from_this())) {
+        if (!is_type<opset1::FakeQuantize>(child)) {
             updatePrecision = true;
             break;
         }
