+++ inference-engine/src/transformations/src/transformations/low_precision/layer_transformation.cpp
@@ -15,6 +15,7 @@
 #include <utility>
 #include <unordered_set>
 #include <vector>
+#include <queue>
 
 namespace ngraph {
 namespace pass {
@@ -345,6 +346,42 @@ void LayerTransformation::fillAvailablePrecisions(std::shared_ptr<Node> layer, s
         }
     }
 }
+
+
+
+std::vector<std::shared_ptr<Node>> LayerTransformation::getChildrenRecursivelyExceptPrecisionPreserved(
+        const std::shared_ptr<Node>& op) const noexcept {
+    std::queue<std::shared_ptr<Node>> notHandledChildren;
+
+    for (const auto& output : op->outputs()) {
+        for (const auto& input : output.get_target_inputs()) {
+            std::shared_ptr<Node> child = input.get_node()->shared_from_this();
+            notHandledChildren.emplace(child);
+        }
+    }
+
+    std::vector<std::shared_ptr<Node>> resultChildren;
+
+    while (!notHandledChildren.empty()) {
+        const std::shared_ptr<ngraph::Node> operation = notHandledChildren.front();
+        notHandledChildren.pop();
+
+        if (!this->isPrecisionPreserved(operation)) {
+            resultChildren.push_back(operation);
+            continue;
+        }
+
+        for (const auto& output : operation->outputs()) {
+            for (const auto& input : output.get_target_inputs()) {
+                std::shared_ptr<Node> child = input.get_node()->shared_from_this();
+                notHandledChildren.emplace(child);
+            }
+        }
+    }
+
+    return resultChildren;
+}
+
 
 std::shared_ptr<ngraph::Node> LayerTransformation::separateInStandaloneBranch(std::shared_ptr<ngraph::Node> node) const {
     FakeQuantizeDequantization dequantization = NetworkHelper::getDequantization(node);
